+++ test/models/feed_profile_test.rb
@@ -24,6 +24,7 @@ class FeedProfileTest < ActiveSupport::TestCase
       "telegram",
       "theycantalk",
       "tomorrows",
+      "tumblr",
       "twitter",
       "webhook",
       "wordpress",